requirements.txt
# ── Terrascope ML Engine Dependencies ────────────────────────

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Validation & config
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
psycopg2-binary==2.9.9

# Cache
redis==5.0.7

# Geospatial
rasterio==1.3.10
geopandas==0.14.4
shapely==2.0.4
h3==3.7.7

# ML
scikit-learn==1.5.1
xgboost==2.0.3
joblib==1.4.2

# Data
numpy==1.26.4
pandas==2.2.2

# HTTP & APIs
requests==2.32.3

# Environment
python-dotenv==1.0.1
